+++ src/main/java/com/gestion/tramite/repositorio/ContactoRepositorio.java
@@ -9,10 +9,10 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ContactoRepositorio extends JpaRepository<Contacto, Integer> {
+public interface ContactoRepositorio extends JpaRepository<Contacto, Long> {
 //Integer id
     @Query("select c from Contacto c where c.persona.id = :idPersona")
-    public List<Contacto> getContactos(@Param("idPersona") Integer idPersona);
+    public List<Contacto> getContactos(@Param("idPersona") Long idPersona);
