+++ src/main/java/com/gestion/tramite/repositorio/DocumentoRepositorio.java
@@ -3,6 +3,6 @@ package com.gestion.tramite.repositorio;
 import com.gestion.tramite.entidad.Documento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentoRepositorio extends JpaRepository<Documento, Integer> {
+public interface DocumentoRepositorio extends JpaRepository<Documento, Long> {
 
 }
