+++ src/main/java/com/gestion/tramite/repositorio/GestionTramiteRepositorio.java
@@ -5,7 +5,7 @@ import com.gestion.tramite.entidad.GestionTramite;
 import com.gestion.tramite.entidad.Persona;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GestionTramiteRepositorio extends JpaRepository<GestionTramite, Integer> {
+public interface GestionTramiteRepositorio extends JpaRepository<GestionTramite, Long> {
 
 
 }
