+++ src/main/java/com/gestion/tramite/repositorio/PersonaRepositorio.java
@@ -4,7 +4,7 @@ package com.gestion.tramite.repositorio;
 import com.gestion.tramite.entidad.Persona;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonaRepositorio extends JpaRepository<Persona, Integer> {
+public interface PersonaRepositorio extends JpaRepository<Persona, Long> {
 
 
 }
