+++ src/main/java/com/gestion/tramite/repositorio/TipoTramiteRepositorio.java
@@ -3,7 +3,7 @@ package com.gestion.tramite.repositorio;
 import com.gestion.tramite.entidad.TipoTramite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoTramiteRepositorio extends JpaRepository<TipoTramite, Integer> {
+public interface TipoTramiteRepositorio extends JpaRepository<TipoTramite, Long> {
 
 
 }
