+++ src/main/java/com/gestion/tramite/service/GestionTramiteService.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 public interface GestionTramiteService {
 
     public List<GestionTramite> listAllGestionTramite();
-    public GestionTramite getGestionTramite(Integer id);
+    public GestionTramite getGestionTramite(Long id);
     public GestionTramite createGestionTramite(GestionTramite a1);
     public GestionTramite updateGestionTramite(GestionTramite a1);
-    public GestionTramite deleteGestionTramite(Integer id);
-    public void borrarGestionTramite(Integer id);
+    public GestionTramite deleteGestionTramite(Long id);
+    public void borrarGestionTramite(Long id);
 }
