+++ src/main/java/com/gestion/tramite/service/PersonaService.java
@@ -11,10 +11,10 @@ import java.util.List;
 public interface PersonaService {
 
     public List<Persona> listAllCliente();
-    public Persona getPersona(Integer id);
+    public Persona getPersona(Long id);
     public Persona createPersona(Persona a1);
     public Persona updatePersona(Persona a1);
-    public Persona deletePersona(Integer id);
-    public void borrarPersona(Integer id);
+    public Persona deletePersona(Long id);
+    public void borrarPersona(Long id);
     public Resource load(String filename,String carpeta);
 }
